Replace ADCIRC flux BC with water elevation BC
TO DO:
If there are multiple open boundaries with BC series specified, then
the current way of overriding fort.19 will cause problems and might
fail.
-------------

The changes were needed to allow the Neural network to force ADCIRC
through a water surface elevation BC instead of flux.

It was fairly easy to change from flux to wse in this commit.
Find and replace
- vout, qout --> elev (in most places),
- nvell --> nvdll,
- nvel --> nope,
- ftiminc --> etiminc,
- qtime1, qtime2 --> etime1, etime2,
- qnin1, qnin2 --> esbin1, esbin2, and
- fort20, fort.20 --> fort19, fort.19,
along with a few other changes did the trick.